archsim: Refactor devices to support 64 bit accesses, work more on riscv64 system model

+++ archsim/inc/abi/devices/riscv/SifiveCLINT.h
@@ -29,8 +29,8 @@ namespace archsim
 					virtual ~SifiveCLINT();
 
 					bool Initialise() override;
-					bool Read(uint32_t offset, uint8_t size, uint32_t& data) override;
-					bool Write(uint32_t offset, uint8_t size, uint32_t data) override;
+					bool Read(uint32_t offset, uint8_t size, uint64_t& data) override;
+					bool Write(uint32_t offset, uint8_t size, uint64_t data) override;
 
 				private:
 					archsim::core::thread::ThreadInstance *GetHart(int i);
